sched: separate thread::status for stopped and running threads
TODO: barrier & memory_order review

The current code (obviously) works, but the thread::status is not always
semantically correct: it is impossible to tell from  _detached_state->st
whether a given thread is still running on another CPU or whether it has
schedule out (=stopped).

This information is crucial for thread migration and - honestly - it is
still not clear to me which of the states introduced in this commit have
previously been merged / overlapping.

The waiting path now cleanly distinguishes between a running thread and
a stopped thread, in the various stages of
  waiting* / sending_lock* / waking* / stagemig*
by suffixing states where the thread is still running with _run
and states where the thread has been stopped with _sto.

It can be shown that there is a total order on the 3 pseudo-states
waiting* < sending_lock* < waking* where '<' means 'before'.
This total order allows to omit retry-loops in wake_lock and wake_impl
by performing their cmpxchg-based state transitions in that order.

The transition matrix comment in sched.hh now is valid Graphviz DOT
graph code and can be extracted using the new
`make thread_state.dot` target.

ON SENDING LOCK / WAIT MORPHING

A side note on sending_lock*, because I will use this commit message when
writing my bachelor thesis and probably have forgotten this by then:
The unlocker calls locker->wake_lock (indirectly) to transition the locker
from waiting* to sending_lock*.
The unlocker completes unlocking and calls locker->wake_impl (indirectly)
to transition the locker from sending_lock* to waking*.