stagesched: busy-wait for comp on thread migration path + use cmpxchg
Idea behind this: the time between enqueuing a thread for thread migration
and the completion (thread not running on source anymore) should already
be minimal.
By busy-waiting for the completion to finish, we reveal bugs in case the
migration takes longer than expected.

The use of cmpxchg merely does in one instruction what we did before in
multiple ones.

TODO: consider relaxed memory ordering while busy-waiting?